python_ml_service/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0
PyMySQL>=1.0.0
